Add files via upload
"The infinite monkey theorem states that a monkey hitting keys at random on a typewriter keyboard for an infinite amount of time will almost surely type any given text, such as the complete works of William Shakespeare.", says Wikipedia. Because I don't have a monkey, a typewriter keyboard and an infinite amount of time, I coded a little script in python that allows you to simulate the typing nearly the same way the monkey does.

Important: It only works on mac (apple) and also it isn't the same as a monkey because it generates a new set of random characters every time.